+++ sand-components/src/lib.rs
@@ -73,10 +73,10 @@ pub use predicates::{
     IntRange, ItemPredicate, LocationPredicate, Range,
 };
 pub use registry::{
-    AdvancementId, BiomeId, BlockId, DamageTypeId, DimensionId, DimensionTypeId,
+    AdvancementId, BiomeId, BlockId, DamageTypeId, DensityFunctionId, DimensionId, DimensionTypeId,
     EnchantmentEffectComponentId, EnchantmentId, EntityTypeId, EquipmentModelId, FunctionId,
-    ItemId, LootTableId, PotionRegistryId, RecipeId, SoundEventId, StatusEffectId, StructureId,
-    TagId,
+    ItemId, LootTableId, NoiseId, PotionRegistryId, RecipeId, SoundEventId, StatusEffectId,
+    StructureId, TagId,
 };
 pub use resource_location::{Identifier, PackNamespace, ResourceLocation};
 
@@ -179,5 +179,6 @@ pub use wolf_variant::WolfVariant;
 
 pub use worldgen::biome::BiomeEffects;
 pub use worldgen::{
-    Biome, Dimension, DimensionType, MonsterSpawnLightLevel, NoiseSettings, PlacedFeature,
+    Biome, DensityFunction, DensityFunctionBinaryOp, DensityFunctionExpr, DensityFunctionUnaryOp,
+    Dimension, DimensionType, MonsterSpawnLightLevel, Noise, NoiseSettings, PlacedFeature,
 };